+++ src/ymm/keys.py
@@ -2,6 +2,7 @@
 kCall='$'
 kLog='debug'
 kLast='^'
+kPrior='^^'
 
 DEFAULT_FILE="./ymm.yml"
 VERSION_FILE='version.txt'