+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 
 - kPrior
 - undo version hack
+- always echo action
 
 
 0.3.0 (2022-03-05)