+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 - kPrior
 - undo version hack
 - always echo action
+- store results
+- keyed output
 
 
 0.3.0 (2022-03-05)